+++ src/test/groovy/ru/vyarus/guice/persist/orient/AbstractTest.groovy
@@ -1,11 +1,9 @@
 package ru.vyarus.guice.persist.orient
 
 import com.google.inject.persist.PersistService
-import com.orientechnologies.orient.core.tx.OTransaction
+import com.orientechnologies.orient.core.db.document.ODatabaseDocumentTx
 import com.orientechnologies.orient.object.db.OObjectDatabaseTx
 import ru.vyarus.guice.persist.orient.db.transaction.TransactionManager
-import ru.vyarus.guice.persist.orient.db.transaction.TxConfig
-import ru.vyarus.guice.persist.orient.db.transaction.template.SpecificTxAction
 import ru.vyarus.guice.persist.orient.db.transaction.template.SpecificTxTemplate
 import ru.vyarus.guice.persist.orient.support.Config
 import spock.lang.Specification
@@ -32,11 +30,7 @@ abstract class AbstractTest extends Specification {
     }
 
     void cleanup() {
-        // truncate db
-        template.doInTransaction(new TxConfig(OTransaction.TXTYPE.NOTX), { db ->
-            db.getStorage().clusterInstances.each({ it.delete() });
-            db.getEntityManager().deregisterEntityClasses(Config.MODEL_PKG)
-        } as SpecificTxAction<Void, OObjectDatabaseTx>)
         persist.stop()
+        new ODatabaseDocumentTx(Config.DB).open(Config.USER, Config.PASS).drop()
     }
 }